+++ src/lib/Advertiser.ts
@@ -179,7 +179,7 @@ export class CiaoAdvertiser extends EventEmitter implements Advertiser {
   static computeSetupHash(accessoryInfo: AccessoryInfo): string {
     const hash = crypto.createHash("sha512");
     hash.update(accessoryInfo.setupID + accessoryInfo.username.toUpperCase());
-    return hash.digest().slice(0, 4).toString("base64");
+    return hash.digest().subarray(0, 4).toString("base64");
   }
 
   public static ff(...flags: PairingFeatureFlag[]): number {