+++ src/lib/camera/RecordingManagement.ts
@@ -1022,7 +1022,7 @@ class CameraRecordingStream extends EventEmitter implements DataStreamProtocolHa
             break;
           }
 
-          const data = fragment.slice(offset, offset + maxChunk);
+          const data = fragment.subarray(offset, offset + maxChunk);
           offset += data.length;
 
           // see https://github.com/Supereg/secure-video-specification#42-binary-data